README.md
This projecy looks at the analysis of a dataset of chemical properties of wines originating from the same region in Italy. After removal of group labels, some clustering techniques are used to find the natural groupings and also statistical learning techniques such as LASSO and Ridge Regression. This will help analyze the pattern of the data and study the efficiency of the clustering methods and also to improve model generalization by reducing overfitting.